1104 refactor family income section (#1124)
* 1104: Fixup family determination model + validations

* Attempts to create a wrapping component "WithNewDetermination" to handle
creating the new (i.e. with id = 0) determination if it does not exist.
(WithNewDetermination lives in FamilyIncome_New/Fields)

HOWEVER: The update data call made there does not seem to ...work?
happen? something. Because even when the associated fields are
successfully forced to re-render after the call to updateData by
WithNewDetermination, the form data object still does NOT have the new
det. The new det finally appears after a set of renders triggered by the
call to updateData from updating a form field. The result is two new
determinations are added: one with the updated form field value and
nothing else, and one with the id value and nothing else. Subsequent
calls to updateData (via form field updates) will affect the det with id
=0

is a fix here to abandon all the `find` stuff all together? and have a
fixed index? since you can only add one new det per render, that should
be fine... It wont' FIX the problem above, but will work around it by
not caring whether or not there's an id, and then just always updating
the value at the fixed index when form fields are updated

* Fully works huzzah

* delete old!

* Rename new

* Keep cleaning up

* Use catchall error

* EnrollmentEdit -> ENrollmentUpdate and comments

* Tests & formatting

* PR comments, more tests, reformat

* Formatting + enrollment history -> enrollment timeline + enrollmentDetail w/out deepNonUndefineable

* Rebase conflicts

* Align with ChildInfo as much as poss

* Test + accessibility fixes, formatting, sonarcloud

* new PR comments

* Style tweak

* Snapshots